# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API для управления онлайн-подписками пользователей",
        "title": "Subscriptions API",
        "contact": {
            "name": "Artem",
            "email": "disaer21@yandex.ru"
        },
        "version": "2.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/subscriptions": {
            "post": {
                "description": "Создает новую подписку для пользователя",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subscriptions"
                ],
                "summary": "Создать подписку",
                "parameters": [
                    {
                        "description": "Данные подписки",
                        "name": "subscription",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.CreateSubRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Подписка успешно создана",
                        "schema": {
                            "$ref": "#/definitions/api.CreateSubResponse"
                        }
                    },
                    "400": {
                        "description": "Ошибка валидации",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Подписка уже существует",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/users/{user_id}/subscriptions": {
            "get": {
                "description": "Возвращает список подписок для указанного user_id. Можно фильтровать по статусу и пагинировать.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subscriptions"
                ],
                "summary": "Получить подписки пользователя",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID пользователя",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "enum": [
                            "active",
                            "archived"
                        ],
                        "type": "string",
                        "default": "active",
                        "description": "Статус подписки",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Номер страницы для пагинации",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список подписок",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/api.SubResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Некорректный UUID пользователя, service_name, статус или номер страницы",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/users/{user_id}/subscriptions/{service_name}": {
            "get": {
                "description": "Возвращает список подписок для указанного user_id. Можно фильтровать по статусу и пагинировать.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subscriptions"
                ],
                "summary": "Получить подписки пользователя",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID пользователя",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "enum": [
                            "active",
                            "archived"
                        ],
                        "type": "string",
                        "default": "active",
                        "description": "Статус подписки",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Номер страницы для пагинации",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список подписок",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/api.SubResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Некорректный UUID пользователя, service_name, статус или номер страницы",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Обновляет уровень и/или дату окончания подписки пользователя",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subscriptions"
                ],
                "summary": "Обновить подписку",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID пользователя",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Название сервиса",
                        "name": "service_name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Новые данные подписки",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.UpdateSubRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Сообщение об обновлении подписки",
                        "schema": {
                            "$ref": "#/definitions/api.UpdateSubResponse"
                        }
                    },
                    "400": {
                        "description": "Некорректные данные запроса",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Подписка не найдена",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Удаляет подписку пользователя по дате начала действия",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subscriptions"
                ],
                "summary": "Удалить подписку",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID пользователя",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Название сервиса",
                        "name": "service_name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Дата начала подписки для удаления",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.DeleteSubRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Подписка успешно удалена",
                        "schema": {
                            "$ref": "#/definitions/api.DeleteSubResponse"
                        }
                    },
                    "400": {
                        "description": "Некорректные данные запроса",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Подписка не найдена",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/users/{user_id}/subscriptions/{service_name}/total": {
            "post": {
                "description": "Рассчитывает общую стоимость подписки за период",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subscriptions"
                ],
                "summary": "Подсчитать общую стоимость подписки",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID пользователя",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Название сервиса",
                        "name": "service_name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Период total_from / total_to",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.TotalCostRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Сообщение с суммой подписки",
                        "schema": {
                            "$ref": "#/definitions/api.TotalCostResponse"
                        }
                    },
                    "400": {
                        "description": "Некорректные данные запроса",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.CreateSubRequest": {
            "type": "object",
            "properties": {
                "end_date": {
                    "type": "string",
                    "example": "12-2025"
                },
                "price": {
                    "type": "integer",
                    "example": 100
                },
                "service_name": {
                    "type": "string",
                    "example": "Yandex Plus"
                },
                "start_date": {
                    "type": "string",
                    "example": "07-2025"
                },
                "user_id": {
                    "type": "string",
                    "example": "60601fee-2bf1-4721-ae6f-7636e79a0cba"
                }
            }
        },
        "api.CreateSubResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Подписка успешно создана"
                }
            }
        },
        "api.DeleteSubRequest": {
            "type": "object",
            "properties": {
                "start_date": {
                    "type": "string",
                    "example": "07-2025"
                }
            }
        },
        "api.DeleteSubResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Подписка успешно удалена"
                }
            }
        },
        "api.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Некорректный формат идентификатора пользователя"
                }
            }
        },
        "api.SubResponse": {
            "type": "object",
            "properties": {
                "end_date": {
                    "type": "string",
                    "example": "12-2025"
                },
                "price": {
                    "type": "integer",
                    "example": 100
                },
                "service_name": {
                    "type": "string",
                    "example": "Yandex Plus"
                },
                "start_date": {
                    "type": "string",
                    "example": "07-2025"
                }
            }
        },
        "api.TotalCostRequest": {
            "type": "object",
            "properties": {
                "total_from": {
                    "type": "string",
                    "example": "07-2025"
                },
                "total_to": {
                    "type": "string",
                    "example": "09-2025"
                }
            }
        },
        "api.TotalCostResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Общая стоимость подписки Yandex Plus за указанный период составила: 300"
                }
            }
        },
        "api.UpdateSubRequest": {
            "type": "object",
            "properties": {
                "new_end_date": {
                    "type": "string",
                    "example": "12-2025"
                },
                "new_price": {
                    "type": "integer",
                    "example": 100
                }
            }
        },
        "api.UpdateSubResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Уровень подписки повышен и уже действует. Дата окончания подписки изменена"
                }
            }
        }
    }
}